Move Get and TestDBGet to their own files
